refactor: move handler to separate class
